+++ test/manual/Pixelate-test.ts
@@ -1,6 +1,5 @@
-import { assert } from 'chai';
-
 import { addStage, Konva, loadImage } from '../unit/test-utils';
+import { cloneAndCompareLayer } from '../unit/test-utils';
 
 describe('Pixelate', function () {
   // ======================================================
@@ -41,5 +40,26 @@ describe('Pixelate', function () {
 
       done();
     });
+  });
+
+  it('make sure we have no extra transparent pixels', function (done) {
+    var stage = addStage();
+    var layer = new Konva.Layer();
+    stage.add(layer);
+
+    Konva.Image.fromURL(
+      'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAGUAAABmCAYAAADS6F9hAAAAAXNSR0IArs4c6QAAAXJJREFUeF7t1cEJADAMw8B2/6Fd6BT3UCYQEiZ3205HGbhFoXp8mKJ4TYoCNilKUUQDIFM/pSigARCppRQFNAAitZSigAZApJZSFNAAiNRSigIaAJFaSlFAAyBSSykKaABEailFAQ2ASC2lKKABEKmlFAU0ACK1lKKABkCkllIU0ACI1FKKAhoAkVpKUUADIFJLKQpoAERqKUUBDYBILaUooAEQqaUUBTQAIrWUooAGQKSWUhTQAIjUUooCGgCRWkpRQAMgUkspCmgARGopRQENgEgtpSigARCppRQFNAAitZSigAZApJZSFNAAiNRSigIaAJFaSlFAAyBSSykKaABEailFAQ2ASC2lKKABEKmlFAU0ACK1lKKABkCkllIU0ACI1FKKAhoAkVpKUUADIFJLKQpoAERqKUUBDYBILaUooAEQqaUUBTQAIrWUooAGQKSWUhTQAIjUUooCGgCRWkpRQAMgUkspCmgARGopRQENgEgPgGOW3jCsp3sAAAAASUVORK5CYII=',
+      function (image) {
+        layer.add(image);
+
+        image.cache();
+        image.filters([Konva.Filters.Pixelate]);
+        image.pixelSize(4);
+        layer.draw();
+        cloneAndCompareLayer(layer);
+
+        done();
+      }
+    );
   });
 });