+++ test/unit/Text-test.ts
@@ -1654,7 +1654,7 @@ describe('Text', function () {
 
     assert.equal(layer.getContext().getTrace(), trace);
   });
-  
+
   it('sets ltr text direction', function () {
     var stage = addStage();
     var layer = new Konva.Layer();
@@ -1673,9 +1673,8 @@ describe('Text', function () {
 
     assert.equal(layer.getContext().getTrace(), trace);
   });
-  
-  
- it('sets rtl text direction', function () {
+
+  it('sets rtl text direction', function () {
     var stage = addStage();
     var layer = new Konva.Layer();
 
@@ -1693,7 +1692,7 @@ describe('Text', function () {
 
     assert.equal(layer.getContext().getTrace(), trace);
   });
-  
+
   it('sets rtl text direction with letterSpacing', function () {
     var stage = addStage();
     var layer = new Konva.Layer();
@@ -1710,6 +1709,24 @@ describe('Text', function () {
 
     var trace =
       'clearRect(0,0,578,200);clearRect(0,0,578,200);save();transform(1,0,0,1,0,0);direction=rtl;font=normal normal 12px Arial;textBaseline=middle;textAlign=left;translate(0,0);save();letterSpacing=2px;fillStyle=black;fillText(rtl text,0,6);restore();restore();';
+
+    assert.equal(layer.getContext().getTrace(), trace);
+  });
+
+  it('try fixed render', () => {
+    Konva._fixTextRendering = true;
+    var stage = addStage();
+    var layer = new Konva.Layer();
+
+    stage.add(layer);
+    var text = new Konva.Text({ text: 'hello', fontSize: 100 });
+
+    layer.add(text);
+    layer.draw();
+    Konva._fixTextRendering = false;
+
+    const trace =
+      'clearRect(0,0,578,200);clearRect(0,0,578,200);save();transform(1,0,0,1,0,0);font=normal normal 100px Arial;textBaseline=alphabetic;textAlign=left;translate(0,0);save();fillStyle=black;fillText(hello,0,85);restore();restore();';
 
     assert.equal(layer.getContext().getTrace(), trace);
   });